requirements.txt
fastapi>=0.110
uvicorn[standard]>=0.23
pydantic>=2.6
python-dotenv>=1.0
langchain>=0.2
langchain-core>=0.2
langchain-google-genai>=1.0.6
google-generativeai>=0.7
httpx>=0.27